+++ leitus/tests/test_ext_file_system.py
@@ -19,7 +19,7 @@
 
 from unittest import mock
 
-from leitus import deep
+from leitus import filesystem
 
 SAMPLE_NAME = "music"
 SAMPLE_COMMAND = ["dumpe2fs", "-h", "/dev/mapper/music"]
@@ -77,17 +77,17 @@ Journal start:            0
 """
 
 
-@mock.patch('leitus.deep.subprocess')
+@mock.patch('leitus.filesystem.subprocess')
 def test_headers_call(mock_subprocess):
-    deep.ExtFileSystem.headers(SAMPLE_NAME)
+    filesystem.ExtFileSystem.headers(SAMPLE_NAME)
 
     mock_subprocess.check_output.assert_has_calls([
         mock.call(SAMPLE_COMMAND)
     ])
 
 
-@mock.patch('leitus.deep.subprocess')
+@mock.patch('leitus.filesystem.subprocess')
 def test_headers_output(mock_subprocess):
     mock_subprocess.check_output.return_value = SAMPLE_OUTPUT
 
-    assert deep.ExtFileSystem.headers(SAMPLE_NAME).raw == SAMPLE_OUTPUT
+    assert filesystem.ExtFileSystem.headers(SAMPLE_NAME).raw == SAMPLE_OUTPUT
